EPI-1756: Refactorings and additions
- Replaced int IDs with long IDs
- refined LiquiBase schema and data initialisation

+++ TOMS/Backend/TOMS-Common/src/main/java/com/matthey/pmm/toms/transport/AttributeCalculationTo.java
@@ -15,7 +15,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 @JsonDeserialize(as = ImmutableAttributeCalculationTo.class)
 @JsonRootName (value = "attributeCalculation")
 public abstract class AttributeCalculationTo {
-    public abstract int id();
+    public abstract long id();
     
     @Auxiliary
     public abstract String className();
