+++ TOMS/Backend/TOMS-MockService/src/main/java/com/matthey/pmm/toms/service/mock/MockCreditCheckService.java
@@ -7,7 +7,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -36,12 +36,12 @@ import io.swagger.annotations.ApiParam;
 
 @RestController
 public class MockCreditCheckService implements TomsCreditCheckService {
-	public static final AtomicInteger ID_COUNTER_CREDIT_LIMIT_CHECK = new AtomicInteger(10000);
+	public static final AtomicLong ID_COUNTER_CREDIT_LIMIT_CHECK = new AtomicLong(10000);
 	public static final List<CreditCheckTo> CUSTOM_CREDIT_LIMIT_CHECKS = new CopyOnWriteArrayList<>();
     
     @ApiOperation("Retrieval of the Credit Check data for a Limit Order")
     public Set<CreditCheckTo> getCreditCheckLimitOrders (
-    		@ApiParam(value = "The order ID of the order the Credit Check object is to be retrieved from", example = "1") @PathVariable int limitOrderId) {
+    		@ApiParam(value = "The order ID of the order the Credit Check object is to be retrieved from", example = "1") @PathVariable long limitOrderId) {
     	Stream<OrderTo> allDataSources = Stream.concat(TestLimitOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
     	LimitOrderTo limitOrder = SharedMockLogic.validateOrderId(this.getClass(), "getCreditCheckLimitOrders", "limitOrderId", limitOrderId, allDataSources, LimitOrderTo.class);
     	
@@ -57,8 +57,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     }
     
     @ApiOperation("Creation of a new Credit Check for a Limit Order")
-    public int postLimitOrderCreditCheck (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable int limitOrderId,
+    public long postLimitOrderCreditCheck (
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable long limitOrderId,
     		@ApiParam(value = "The new Credit Check. ID has to be -1. The actual assigned ID is going to be returned", example = "", required = true) @RequestBody(required=true) CreditCheckTo newCreditCheck) {
        	Stream<OrderTo> allDataSources = Stream.concat(TestLimitOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
        	LimitOrderTo limitOrder = SharedMockLogic.validateOrderId (this.getClass(), "postLimitOrderCreditCheck", "limitOrderId", limitOrderId, allDataSources, LimitOrderTo.class);
@@ -73,7 +73,7 @@ public class MockCreditCheckService implements TomsCreditCheckService {
 			.filter(x-> x.getEntity().id() == limitOrderId)
 			.collect(Collectors.toList());
 		
-		Set<Integer> newCreditCheckIds = new HashSet<>(limitOrder.creditChecksIds());
+		Set<Long> newCreditCheckIds = new HashSet<>(limitOrder.creditChecksIds());
 		newCreditCheckIds.add(withId.id());
 		
 		SimpleDateFormat sdfDateTime = new SimpleDateFormat (TomsService.DATE_TIME_FORMAT);
@@ -94,8 +94,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
 
     @ApiOperation("Update of the Credit Check for a Limit Order")
     public void updateLimitOrderCreditCheck (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable int limitOrderId,
-    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable int creditCheckId,
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable long limitOrderId,
+    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable long creditCheckId,
     		@ApiParam(value = "The updated Credit Check. ID has to be matching the ID of the existing Credit Check.", example = "", required = true) @RequestBody(required=true) CreditCheckTo existingCreditCheck) {
 		Stream<OrderTo> allDataSources = Stream.concat(TestLimitOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
 		LimitOrderTo limitOrder = SharedMockLogic.validateOrderId (this.getClass(), "updateLimitOrderCreditCheck", "limitOrderId", limitOrderId, allDataSources, LimitOrderTo.class);
@@ -140,8 +140,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     
     @ApiOperation("Retrieval of the Credit Check data for a Limit Order")
     public CreditCheckTo getCreditCheckLimitOrder (
-    		@ApiParam(value = "The order ID of the limit order the Credit Check is to be retrieved from", example = "1") @PathVariable int limitOrderId,
-    		@ApiParam(value = "The ID of the Credit Check to retrieve ", example = "1") @PathVariable int creditCheck) {
+    		@ApiParam(value = "The order ID of the limit order the Credit Check is to be retrieved from", example = "1") @PathVariable long limitOrderId,
+    		@ApiParam(value = "The ID of the Credit Check to retrieve ", example = "1") @PathVariable long creditCheck) {
     	Stream<OrderTo> allDataSources = Stream.concat(TestLimitOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
     	LimitOrderTo limitOrder = SharedMockLogic.validateOrderId(this.getClass(), "getCreditCheckLimitOrder", "limitOrderId", limitOrderId, allDataSources, LimitOrderTo.class);
     	
@@ -161,8 +161,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     }
     
     @ApiOperation("Creation of a new Credit Check for a Reference Order")
-    public int postReferenceOrderCreditCheck (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable int referenceOrderId,
+    public long postReferenceOrderCreditCheck (
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable long referenceOrderId,
     		@ApiParam(value = "The new Credit Check . ID has to be -1. The actual assigned ID is going to be returned", example = "", required = true) @RequestBody(required=true) CreditCheckTo newCreditCheck) {
        	Stream<OrderTo> allDataSources = Stream.concat(TestReferenceOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
        	ReferenceOrderTo referenceOrder = SharedMockLogic.validateReferenceOrderId (this.getClass(), "postReferenceOrderCreditCheck", "referenceOrderId", referenceOrderId, allDataSources);
@@ -177,7 +177,7 @@ public class MockCreditCheckService implements TomsCreditCheckService {
 			.filter(x-> x.getEntity().id() == referenceOrderId)
 			.collect(Collectors.toList());
 		
-		Set<Integer> newCreditCheckIds = new HashSet<>(referenceOrder.creditChecksIds());
+		Set<Long> newCreditCheckIds = new HashSet<>(referenceOrder.creditChecksIds());
 		newCreditCheckIds.add(withId.id());
 		
 		SimpleDateFormat sdfDateTime = new SimpleDateFormat (TomsService.DATE_TIME_FORMAT);
@@ -197,8 +197,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     
     @ApiOperation("Update of the Credit Check for a Reference Order")
     public void updateReferenceOrderCreditCheck (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable int referenceOrderId,
-    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable int creditCheckId,
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be posted for ", example = "1") @PathVariable long referenceOrderId,
+    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable long creditCheckId,
     		@ApiParam(value = "The updated Credit Check. ID has to be matching the ID of the existing Credit Check.", example = "", required = true) @RequestBody(required=true) CreditCheckTo existingCreditCheck) {
 		Stream<OrderTo> allDataSources = Stream.concat(TestReferenceOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
     	ReferenceOrderTo referenceOrder = SharedMockLogic.validateReferenceOrderId (this.getClass(), "updateReferenceOrderCreditCheck", "referenceOrderId", referenceOrderId, allDataSources);
@@ -243,7 +243,7 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     
     @ApiOperation("Retrieval of the Credit Check data for a Reference Order")
     public Set<CreditCheckTo> getCreditChecksReferenceOrders (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be retrieved from", example = "1") @PathVariable int referenceOrderId) {
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be retrieved from", example = "1") @PathVariable long referenceOrderId) {
     	Stream<OrderTo> allDataSources = Stream.concat(TestReferenceOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
     	ReferenceOrderTo referenceOrder = SharedMockLogic.validateReferenceOrderId(this.getClass(), "getCreditChecksReferenceOrders", "limitOrderId", referenceOrderId, allDataSources);
     	
@@ -260,8 +260,8 @@ public class MockCreditCheckService implements TomsCreditCheckService {
     
     @ApiOperation("Retrieval of the Credit Check data for a Reference Order")
     public CreditCheckTo getCreditChecksReferenceOrder (
-    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be retrieved from", example = "1") @PathVariable int referenceOrderId,
-    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable int creditCheck) {
+    		@ApiParam(value = "The order ID of the reference order the Credit Check is to be retrieved from", example = "1") @PathVariable long referenceOrderId,
+    		@ApiParam(value = "The ID of the Credit Check to update ", example = "1") @PathVariable long creditCheck) {
     	Stream<OrderTo> allDataSources = Stream.concat(TestReferenceOrder.asList().stream(), MockOrderController.CUSTOM_ORDERS.stream());
     	ReferenceOrderTo referenceOrder = SharedMockLogic.validateReferenceOrderId(this.getClass(), "getCreditChecksReferenceOrder", "referenceOrderId", referenceOrderId, allDataSources);
