+++ TOMS/Backend/TOMS-MockService/src/main/java/com/matthey/pmm/toms/service/mock/MockIndexController.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.matthey.pmm.toms.enums.DefaultReferenceType;
+import com.matthey.pmm.toms.enums.v1.DefaultReferenceType;
 import com.matthey.pmm.toms.service.TomsIndexService;
 import com.matthey.pmm.toms.service.TomsService;
 import com.matthey.pmm.toms.service.mock.testdata.TestIndex;
@@ -23,7 +23,7 @@ import io.swagger.annotations.ApiParam;
 public class MockIndexController implements TomsIndexService {
     @ApiOperation("Retrieval of Index Data")
 	public Set<IndexTo> getIndexes (
-			@ApiParam(value = "One of the currencies of the indexes must have the provided ID. Null or 0 = all indexes", example = "42", required = false) @RequestParam(required=false) Integer currencyRefId) {
+			@ApiParam(value = "One of the currencies of the indexes must have the provided ID. Null or 0 = all indexes", example = "42", required = false) @RequestParam(required=false) Long currencyRefId) {
 		if (TomsService.verifyDefaultReference (currencyRefId,
 				Arrays.asList(DefaultReferenceType.CCY_CURRENCY, DefaultReferenceType.CCY_METAL),
 				this.getClass(), "getIndexes", "currencyRefId", false)) {
