+++ TOMS/Backend/TOMS-TestData/src/main/java/com/matthey/pmm/toms/service/mock/testdata/TestOrderComment.java
@@ -19,7 +19,7 @@ public enum TestOrderComment {
 	
 	private OrderCommentTo comment;
 	
-	private TestOrderComment (int id, String commentText, String createdAt, TestUser createdBy,
+	private TestOrderComment (long id, String commentText, String createdAt, TestUser createdBy,
 			String lastUpdate, TestUser updatedBy) {
 		comment = ImmutableOrderCommentTo.builder()
 				.id(id)
